Move interface from sliderbanner to SliderSettings.ts in types folder
